# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "\u5947\u8d27API\u6587\u6863",
        "description": "\u9879\u76ee\u4ee3\u53f7Tiny(\u5c0f\u5c0f)",
        "termsOfService": "https://github.com/powrlaw",
        "contact": {
            "name": "xieyi",
            "url": "http://github.com/xieyi64",
            "email": "dev001@powerlaw.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "xieyi.api.vqiho.com",
    "basePath": "/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/xml",
        "application/json"
    ],
    "produces": [
        "application/xml",
        "application/json"
    ],
    "paths": {
...
        "/goods": {
            "get": {
                "tags": [
                    "\u5546\u54c1"
                ],
                "summary": "\u5217\u8868",
                "description": "\u6682\u65e0",
                "operationId": "good.list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "\u8d77\u59cb\u4f4d\u7f6e",
                        "type": "integer",
                        "default": "0"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "\u83b7\u53d6\u6570\u91cf",
                        "type": "integer",
                        "default": "20"
                    },
                    {
                        "name": "updown",
                        "in": "query",
                        "description": "\u65b9\u5411(1:\u6b63\u5e8f,-1:\u9006\u5e8f)",
                        "type": "integer",
                        "default": "-1"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "\u6392\u5e8f\u5b57\u6bb5(name,-id)",
                        "type": "string"
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "\u67e5\u8be2\u5173\u952e\u8bcd",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/goods/{good_id}": {
            "get": {
                "tags": [
                    "\u5546\u54c1"
                ],
                "summary": "\u5217\u8868",
                "description": "\u6682\u65e0",
                "operationId": "good.show",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "good_id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/goods/favors": {
            "get": {
                "tags": [
                    "\u5546\u54c1"
                ],
                "summary": "\u6536\u85cf\u5217\u8868",
                "description": "\u6682\u65e0",
                "operationId": "good.favor.list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "\u8d77\u59cb\u4f4d\u7f6e",
                        "type": "integer",
                        "default": "0"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "\u83b7\u53d6\u6570\u91cf",
                        "type": "integer",
                        "default": "20"
                    },
                    {
                        "name": "updown",
                        "in": "query",
                        "description": "\u65b9\u5411(1:\u6b63\u5e8f,-1:\u9006\u5e8f)",
                        "type": "integer",
                        "default": "-1"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "\u6392\u5e8f\u5b57\u6bb5(name,-id)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/goods/{good_id}/favors": {
            "post": {
                "tags": [
                    "\u5546\u54c1"
                ],
                "summary": "\u6536\u85cf",
                "description": "\u6682\u65e0",
                "operationId": "good.favor.store",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "good_id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "\u5546\u54c1"
                ],
                "summary": "\u53d6\u6d88\u6536\u85cf",
                "description": "\u6682\u65e0",
                "operationId": "good.favor.destroy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "good_id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
        "/videos/{video_id}/favors": {
            "post": {
                "tags": [
                    "\u89c6\u9891"
                ],
                "summary": "\u6536\u85cf",
                "description": "\u6682\u65e0",
                "operationId": "video.favor.store",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "video_id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "\u89c6\u9891"
                ],
                "summary": "\u53d6\u6d88\u6536\u85cf",
                "description": "\u6682\u65e0",
                "operationId": "video.favor.destroy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "video_id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        }
    },
...
}